requirements.txt
Pillow==9.2.0
argparse==1.4.0